pyproject.toml
[tool.poetry]
name = "discord-gemini-bot"
version = "0.1.0"
description = ""
authors = ["agis09 <agis09m@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord = "^2.3.2"
python-dotenv = "^1.0.1"
google-generativeai = "^0.5.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
